# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface AuthInput {
  email: string;
  password: string;
  name?: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  photo?: string | null;
  language?: string;
  savedDestinations?: string[];
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface ProfileUpdate {
  name?: string;
  photo?: string;
  language?: string;
  savedDestinations?: string[];
}

export interface Trip {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverPhoto?: string | null;
  startDate: string;
  endDate: string;
  isPublic: boolean;
  /** @nullable */
  shareCode?: string | null;
  /** @nullable */
  totalBudget?: number | null;
  stopCount: number;
  createdAt: string;
}

export interface StopActivity {
  id: number;
  stopId: number;
  /** @nullable */
  activityCatalogId?: number | null;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  cost?: number | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  scheduledTime?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Stop {
  id: number;
  tripId: number;
  cityId: number;
  cityName: string;
  country: string;
  /** @nullable */
  cityImageUrl?: string | null;
  startDate: string;
  endDate: string;
  order: number;
  activities?: StopActivity[];
}

export interface TripDetail {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverPhoto?: string | null;
  startDate: string;
  endDate: string;
  isPublic: boolean;
  /** @nullable */
  shareCode?: string | null;
  /** @nullable */
  totalBudget?: number | null;
  createdAt: string;
  stops: Stop[];
}

export interface TripInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  startDate: string;
  endDate: string;
  coverPhoto?: string;
  totalBudget?: number;
}

export interface TripUpdate {
  name?: string;
  description?: string;
  startDate?: string;
  endDate?: string;
  coverPhoto?: string;
  totalBudget?: number;
}

export interface TripShareToggle {
  isPublic: boolean;
}

export interface City {
  id: number;
  name: string;
  country: string;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  costIndex?: number | null;
  /** @nullable */
  popularity?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface StopInput {
  cityId: number;
  startDate: string;
  endDate: string;
}

export interface StopUpdate {
  cityId?: number;
  startDate?: string;
  endDate?: string;
}

export interface StopReorder {
  order: number[];
}

export interface ActivityCatalogItem {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  cost?: number | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  cityId?: number | null;
  /** @nullable */
  cityName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface StopActivityInput {
  activityCatalogId?: number;
  name: string;
  category: string;
  description?: string;
  cost?: number;
  duration?: number;
  scheduledTime?: string;
  imageUrl?: string;
}

export interface BudgetCategory {
  category: string;
  amount: number;
}

export interface StopBudget {
  stopId: number;
  cityName: string;
  startDate: string;
  endDate: string;
  activityCost: number;
  days: number;
}

export interface BudgetBreakdown {
  tripId: number;
  /** @nullable */
  totalBudget?: number | null;
  totalEstimated: number;
  perDayAverage: number;
  /** @nullable */
  remainingBudget?: number | null;
  isOverBudget?: boolean;
  byCategory?: BudgetCategory[];
  stops: StopBudget[];
}

export interface PackingItem {
  id: number;
  tripId: number;
  name: string;
  category: string;
  isPacked: boolean;
  createdAt: string;
}

export interface PackingItemInput {
  /** @minLength 1 */
  name: string;
  category: string;
}

export interface PackingItemUpdate {
  name?: string;
  category?: string;
  isPacked?: boolean;
}

export interface TripNote {
  id: number;
  tripId: number;
  /** @nullable */
  stopId?: number | null;
  content: string;
  createdAt: string;
  updatedAt?: string;
}

export interface TripNoteInput {
  /** @minLength 1 */
  content: string;
  stopId?: number;
}

export interface TripNoteUpdate {
  content?: string;
}

export interface DashboardSummary {
  upcomingTrips: number;
  totalTrips: number;
  totalDestinations: number;
  totalBudgetSpent?: number;
  recentTrips: Trip[];
  popularCities: City[];
}

export type AdminStatsTopCitiesItem = {
  cityName: string;
  count: number;
};

export type AdminStatsRecentActivityItem = {
  description: string;
  createdAt: string;
};

export interface AdminStats {
  totalUsers: number;
  totalTrips: number;
  totalStops: number;
  publicTrips?: number;
  topCities: AdminStatsTopCitiesItem[];
  recentActivity: AdminStatsRecentActivityItem[];
}

export type SearchCitiesParams = {
  q?: string;
  country?: string;
};

export type SearchActivityCatalogParams = {
  q?: string;
  category?: string;
  cityId?: number;
  maxCost?: number;
};