# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Traveloop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Dashboard
...
  /stops/{stopId}/activities:
    get:
      operationId: listStopActivities
      tags: [activities]
      summary: List activities for a stop
      parameters:
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StopActivity"
    post:
      operationId: addStopActivity
      tags: [activities]
      summary: Add an activity to a stop
      parameters:
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopActivity"
  /stops/{stopId}/activities/{activityId}:
    delete:
      operationId: removeStopActivity
      tags: [activities]
      summary: Remove an activity from a stop
      parameters:
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # City search
...
  /activity-catalog:
    get:
      operationId: searchActivityCatalog
      tags: [activityCatalog]
      summary: Search activity catalog
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: cityId
          in: query
          schema:
            type: integer
        - name: maxCost
          in: query
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityCatalogItem"

  # Budget
  /trips/{tripId}/budget:
    get:
      operationId: getTripBudget
      tags: [budget]
      summary: Get budget breakdown for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetBreakdown"

  # Packing
...
  /trips/{tripId}/notes:
    get:
      operationId: listTripNotes
      tags: [notes]
      summary: List notes for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripNote"
    post:
      operationId: createTripNote
      tags: [notes]
      summary: Create a note
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripNote"
  /trips/{tripId}/notes/{noteId}:
    patch:
      operationId: updateTripNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripNoteUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripNote"
    delete:
      operationId: deleteTripNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Profile
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    AuthInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        photo:
          type: ["string", "null"]
        language:
          type: string
        savedDestinations:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    ActivityCatalogItem:
      type: object
      required: [id, name, category]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
        duration:
          type: ["number", "null"]
        cityId:
          type: ["integer", "null"]
        cityName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    StopActivity:
      type: object
      required: [id, stopId, name, category]
      properties:
        id:
          type: integer
        stopId:
          type: integer
        activityCatalogId:
          type: ["integer", "null"]
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
        duration:
          type: ["number", "null"]
        scheduledTime:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    StopActivityInput:
      type: object
      required: [name, category]
      properties:
        activityCatalogId:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        cost:
          type: number
        duration:
          type: number
        scheduledTime:
          type: string
        imageUrl:
          type: string
    BudgetBreakdown:
      type: object
      required: [tripId, totalEstimated, perDayAverage, stops]
      properties:
        tripId:
          type: integer
        totalBudget:
          type: ["number", "null"]
        totalEstimated:
          type: number
        perDayAverage:
          type: number
        remainingBudget:
          type: ["number", "null"]
        isOverBudget:
          type: boolean
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/BudgetCategory"
        stops:
          type: array
          items:
            $ref: "#/components/schemas/StopBudget"
    BudgetCategory:
      type: object
      required: [category, amount]
      properties:
        category:
          type: string
        amount:
          type: number
    StopBudget:
      type: object
      required: [stopId, cityName, startDate, endDate, activityCost, days]
      properties:
        stopId:
          type: integer
        cityName:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        activityCost:
          type: number
        days:
          type: integer
...
    TripNote:
      type: object
      required: [id, tripId, content, createdAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        stopId:
          type: ["integer", "null"]
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    TripNoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        stopId:
          type: integer
    TripNoteUpdate:
      type: object
      properties:
        content:
          type: string
...