# lib/api-zod/src/generated/types/stopBudget.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StopBudget {
  stopId: number;
  cityName: string;
  startDate: string;
  endDate: string;
  activityCost: number;
  days: number;
}